+++ app/src/main/java/evgeniy/ryzhikov/filmsearch/domain/Interactor.kt
@@ -1,0 +1,7 @@
+package evgeniy.ryzhikov.filmsearch.domain
+
+import evgeniy.ryzhikov.filmsearch.data.MainRepository
+
+class Interactor(val repository: MainRepository) {
+    fun getFilmsDB(): List<Film> = repository.filmsDataBase
+}
