+++ app/src/main/java/evgeniy/ryzhikov/filmsearch/view/fragments/SelectionsFragment.kt
@@ -1,4 +1,4 @@
-package evgeniy.ryzhikov.filmsearch.fragments
+package evgeniy.ryzhikov.filmsearch.view.fragments
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
