+++ app/src/main/java/evgeniy/ryzhikov/filmsearch/view/rv_viewholders/FilmViewHolder.kt
@@ -1,13 +1,12 @@
-package evgeniy.ryzhikov.filmsearch.recycler_view
+package evgeniy.ryzhikov.filmsearch.view.rv_viewholders
 
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import evgeniy.ryzhikov.filmsearch.databinding.FilmItemBinding
+import evgeniy.ryzhikov.filmsearch.domain.Film
 
-class FilmViewHolder(private val itemView: View) : RecyclerView.ViewHolder(itemView) {
+class FilmViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     val binding = FilmItemBinding.bind(itemView)
     //В этом методе кладем данные из Film в наши View
     fun bind(film: Film) {
