+++ app/src/main/java/evgeniy/ryzhikov/filmsearch/viewmodel/HomeFragmentViewModel.kt
@@ -1,0 +1,17 @@
+package evgeniy.ryzhikov.filmsearch.viewmodel
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import evgeniy.ryzhikov.filmsearch.App
+import evgeniy.ryzhikov.filmsearch.domain.Film
+import evgeniy.ryzhikov.filmsearch.domain.Interactor
+
+class HomeFragmentViewModel : ViewModel() {
+    val filmsListLiveData = MutableLiveData<List<Film>>()
+    private var interactor = App.instance.interactor
+
+    init {
+        val films = interactor.getFilmsDB()
+        filmsListLiveData.postValue(films)
+    }
+}
