+++ app/src/main/res/layout/activity_greetings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
     <com.airbnb.lottie.LottieAnimationView
         android:id="@+id/greetingsLottieAnim"
