+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
     android:id="@+id/fragmentHome"
     tools:visibility="visible"
     android:background="@drawable/background_fragment_home"
-    tools:context=".fragments.HomeFragment">
+    tools:context=".view.fragments.HomeFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/appBar"
