+++ app/src/main/res/layout/fragment_watch_later.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
@@ -8,6 +7,6 @@
     android:id="@+id/fragmentWatchLater"
     tools:visibility="visible"
     android:background="@drawable/background_fragment_watch_later"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
